Refine deployment workflows and rename release file
Added an additional step in the deploy-main.yml file for checking unmerged changesets and adjusted the condition for the workflow to proceed. Renamed create-release.yml to update-release-pr.yml to better reflect its purpose. The output and variable "deploy" has been renamed to "deployable" for more clarity, and the same changes have been made in the shared.cd.yml file accordingly. The changes aim to automate more finely and improve deployment process.