Update package management tool in GitHub Actions (#44)
* Update package management tool in GitHub Actions

Changed the package management tool from npm to yarn in the .github/workflows/update-release-pr.yml. This change is due to the project's shift to use yarn over npm for more effective dependency management.

* Remove pull request triggers from Changesets workflow

The pull request triggers for the Changesets GitHub action workflow have been removed. This change was made because these triggers were unnecessary, and the workflow should only run on pushes to the main branch instead. This streamlines the workflow process by only running when changes are made to the main branch.